fix validate types tests

+++ sources/backend/accounts/Services/ValidateType/Service.Tests.cs
@@ -6,15 +6,15 @@ namespace Elesse.Accounts.Tests
    {
 
       [Theory]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, 1, 1)]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, null, 1)]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, 1, null)]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.General, 1, 1)]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Investment, 1, 1)]
-      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Service, 1, 1)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, (short)1, (short)1)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, null, (short)1)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Bank, (short)1, null)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.General, (short)1, (short)1)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Investment, (short)1, (short)1)]
+      [InlineData(WARNINGS.DAYS_ONLY_VALID_FOR_CREDIT_CARD_TYPE, enAccountType.Service, (short)1, (short)1)]
       [InlineData(WARNINGS.DAYS_REQUIRED_FOR_CREDIT_CARD_TYPE, enAccountType.CreditCard, null, null)]
-      [InlineData(WARNINGS.DAYS_REQUIRED_FOR_CREDIT_CARD_TYPE, enAccountType.CreditCard, 1, null)]
-      [InlineData(WARNINGS.DAYS_REQUIRED_FOR_CREDIT_CARD_TYPE, enAccountType.CreditCard, null, 1)]
+      [InlineData(WARNINGS.DAYS_REQUIRED_FOR_CREDIT_CARD_TYPE, enAccountType.CreditCard, (short)1, null)]
+      [InlineData(WARNINGS.DAYS_REQUIRED_FOR_CREDIT_CARD_TYPE, enAccountType.CreditCard, null, (short)1)]
       internal async void ValidateType_WithInvalidParameters_MustResultErrorMessage(string warningText, enAccountType type, short? closingDay, short? dueDay)
       {
          var service = new AccountService(null);
@@ -28,7 +28,7 @@ namespace Elesse.Accounts.Tests
       [InlineData(enAccountType.General, null, null)]
       [InlineData(enAccountType.Investment, null, null)]
       [InlineData(enAccountType.Service, null, null)]
-      [InlineData(enAccountType.CreditCard, 1, 1)]
+      [InlineData(enAccountType.CreditCard, (short)1, (short)1)]
       internal async void ValidateUsername_WithValidParameters_MustResultNoErrorMessages(enAccountType type, short? closingDay, short? dueDay)
       {
          var service = new AccountService(null);
